+++ src/components/Auth/AuthForm.tsx
@@ -1,8 +1,10 @@
 import { FormControl, FormLabel, FormHelperText, Input } from "@mui/material";
-import React, { useRef } from "react";
+import React, { useRef, useContext, useState } from "react";
 import classes from "../Layout/AuthLayout/AuthLayout.module.css";
-import useHttp from "../../hooks/use-http";
 import { Formik, Form } from "formik";
+import loginUserRequest from "../lib/authorization";
+import AuthContext from "../../store/auth-context";
+import { useHistory } from "react-router-dom";
 
 interface MyFormValues {
   email: string;
@@ -12,16 +14,37 @@ interface MyFormValues {
 const AuthForm = () => {
   const emailInputRef = useRef<HTMLInputElement>(null);
   const passwordInputRef = useRef<HTMLInputElement>(null);
-
-  const { isLoading, error, loginUserRequest: sendAuthRequest } = useHttp();
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState(null);
+  const authCtx = useContext(AuthContext);
+  const history = useHistory();
 
   const submitHandler = (event: React.FormEvent) => {
     event.preventDefault();
 
     const enteredEmail = emailInputRef.current!.value;
     const enteredPassword = passwordInputRef.current!.value;
+    setIsLoading(true);
 
-    sendAuthRequest(enteredEmail, enteredPassword);
+    loginUserRequest(enteredEmail, enteredPassword)
+      .then((response) => {
+        if (response.status === 200) {
+          console.log(response.data);
+          authCtx.login(response.data.token);
+          history.replace("/dashboard");
+        } else {
+          throw new Error("Authenfication Fail!");
+        }
+      })
+      .catch((error) => {
+        if (error.response.status === 401) {
+          setError(error.response.data.msg);
+        } else {
+          setError(error.response.data.errors[0].msg);
+        }
+        console.log(error);
+        setIsLoading(false);
+      });
   };
 
   const initialValues: MyFormValues = { email: "", password: "" };
