+++ src/types/constants.ts
@@ -1,0 +1,3 @@
+export enum CONSTANTS {
+    URL = 'http://localhost:8080',
+}
